api/frontend/output_handler/requirements.txt
python-socketio[client]